Converted Position Vectors to Floats for smother movement and less casting

+++ include/engine/engine.h
@@ -7,8 +7,8 @@
 #include "math.h"
 
 typedef struct{
-    int* x;
-    int* y;
+    float* x;
+    float* y;
     float* vx;
     float* vy;
     size_t count;
